Hay que arreglar el exportar la base de datos

+++ php/eliminar_factura.php
@@ -1,16 +1,24 @@
 <?php
 require_once "main.php";
-    $pdo = conexion();
 
-    $deuda_id = limpiar_cadena($_POST['deuda_id']);
+$pdo = conexion();
 
-    $stmt = $pdo->prepare("DELETE FROM facturas WHERE id = :id");
-    $stmt->execute([
-    ':id' => $deuda_id
-    ]);
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    $factura_id = limpiar_cadena($_POST['factura_id']);
 
-    	header("Location: index.php?vistas=home")
+    try {
+        $stmt = $pdo->prepare("DELETE FROM facturas WHERE id = :id");
+        $stmt->execute([
+        ':id' => $factura_id
+        ]);
 
-    $pdo = null;
+        header("Location: ../index.php?vistas=Inicio&mensaje=eliminar_factura_exito");
+        exit;
+    } catch (Exception $e) {
 
+        header("Location: ../index.php?vistas=Inicio&mensaje=eliminar_factura_error");
+        exit;
+    }
+}
+$pdo = null;
 ?>
