+++ php/eliminar_persona.php
@@ -2,30 +2,27 @@
 require_once "main.php";
     $pdo = conexion();
 
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $persona_id = limpiar_cadena($_POST['persona_id']);
 
+    try {
     $stmt = $pdo->prepare("DELETE FROM deudas WHERE persona_id = :id");
     $stmt->execute([
-    ':id' => $persona_id
-    ]);
-
-    if(isset($deuda_id)){
-        echo "<div class='mensaje_exito'>
-            <p class='mensaje_exito__p'>El Registro se ha modificado correctamente.</p>
-        </div>";
-
-        header("Location: index.php?vistas=home");
-    }else{
-        echo "<div class='mensaje_exito'>
-            <p class='mensaje_exito__p'>El Registro se ha modificado correctamente.</p>
-        </div>";
-    }
+    ':id' => $persona_id]);
 
     $stmt = $pdo->prepare("DELETE FROM personas WHERE id = :id");
     $stmt->execute([
     ':id' => $persona_id
     ]);
 
-    $pdo = null;
+        header("Location: ../index.php?vistas=Inicio&mensaje=eliminar_persona_exito");
+        exit;
+    } catch (Exception $e) {
+
+        header("Location: ../index.php?vistas=Inicio&mensaje=eliminar_deuda_error");
+        exit;
+    }
+}
+$pdo = null;
 
 ?>
